Move some structs to proper contexts
