+++ service/app/types.go
@@ -58,19 +58,6 @@ type Pack struct {
 	Collectibles   Collectibles       `gorm:"column:collectibles"`          // private
 }
 
-// AddressLocation is a reference to a contract on chain.
-type AddressLocation struct {
-	Name    string             `gorm:"column:name"`
-	Address common.FlowAddress `gorm:"column:address"`
-}
-
-type Collectible struct {
-	FlowID            common.FlowID   // ID of the collectible NFT
-	ContractReference AddressLocation // Reference to the collectible NFT contract
-}
-
-type Collectibles []Collectible
-
 func (Distribution) TableName() string {
 	return "distributions"
 }
